requirements.txt
action-msgs==1.2.1
ament-cmake-test==1.3.10
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-index-python==1.4.0
ament-package==0.14.0
asgiref==3.8.1
builtin-interfaces==1.2.1
Django==5.1.3
djangorestframework==3.15.2
domain-coordinator==0.10.0
gazebo-model-attachment-plugin-msgs==1.0.3
gazebo-msgs==3.7.0
gazebo-video-monitor-interfaces==0.8.1
gazebo-video-monitor-utils==0.8.1
geometry-msgs==4.2.4
lifecycle-msgs==1.2.1
pillow==11.0.0
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
rmw-dds-common==1.6.0
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
sqlparse==0.5.1
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
trajectory-msgs==4.2.4
typing_extensions==4.12.2
unique-identifier-msgs==2.2.1
